requirements.txt
PyQt5>=5.15.0
python-dotenv>=0.19.0 